jack: fix type warning in error message
just casting to int because it's the simplest (%z is not
well-supported)

Noticed-by: avuton on a 64-bit machine

git-svn-id: https://svn.musicpd.org/mpd/trunk@5257 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/audioOutputs/audioOutput_jack.c
@@ -203,8 +203,8 @@ static int jack_initDriver(AudioOutput *audioOutput, ConfigParam *param)
 
 		if (strchr(cp,','))
 			FATAL("Only %d values are supported for '%s' "
-			      "at line %d\n",
-			      ARRAY_SIZE(output_ports), bp->name, bp->line);
+			      "at line %d\n", (int)ARRAY_SIZE(output_ports),
+			      bp->name, bp->line);
 	}
 
 	if ( (bp = getBlockParam(param, "ringbuffer_size")) ) {
